requirements.txt
flask
python-dotenv
flask-wtf
flask_sqlalchemy
flask_migrate
flask-login
email-validator
eventlet
flask-mail
pyjwt
flask-dotenv
flask-moment
flask-bootstrap
langdetect
flask_babel
gevent
HTTPie
flask-httpauth
flask-pagedown==0.4.0
markdown==3.5.1
bleach==6.1.0
flask-socketio
flask-dropzone
flask-avatars
elasticsearch
rq